pyproject.toml
[tool.poetry]
name = "my_observatory_test"
version = "0.1.0"
description = "Automated tests for MyObservatory app"
authors = ["Edwin Law <edwinkhlaw@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.9.6"
Appium-Python-Client = "5.0.0"
pytest = "8.3.5"
selenium = "4.30.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"